requirements.txt
alembic==1.7.4
annotated-types==0.6.0
anyio==3.7.1
attrs==21.2.0
backports.entry-points-selectable==1.1.0
blinker==1.7.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.4
decorator==5.1.1
distlib==0.3.3
distro==1.9.0
dnspython==2.6.1
exceptiongroup==1.2.0
filelock==3.3.1
Flask==3.0.0
Flask-SQLAlchemy==2.5.1
getenv==0.2.0
greenlet==1.1.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.3
jsonschema==4.1.2
linecache2==1.0.0
loguru==0.7.2
Mako==1.1.5
MarkupSafe==2.1.5
mistune==0.8.4
numpy==1.26.4
openai==1.3.7
pinecone-client==2.2.4
pip==24.0
platformdirs==2.4.0
py==1.11.0
pydantic==2.7.0
pydantic_core==2.18.1
pymongo==4.6.1
pyrsistent==0.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0
redis==2.10.6
regex==2024.4.16
requests==2.31.0
retry==0.9.2
scipy==1.11.4
setuptools==63.2.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.26
tavily-python==0.2.8
tiktoken==0.5.1
tqdm==4.66.2
traceback2==1.4.0
typing_extensions==4.11.0
unittest2==1.1.0
urllib3==1.26.7
virtualenv==20.9.0
Werkzeug==3.0.2
win32-setctime==1.1.0
WTForms==2.3.3
